requirements.txt
kafka-python>=2.0.2
pymongo>=4.6.0
python-dotenv>=1.0.0